# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "XSocial Laravel API",
        "description": "✊✊✊",
        "contact": {
            "email": "ngoctoan06011998@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "posts.index",
                "operationId": "e9eaf7df702a94f4443677bd1a36cfe6",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "filter[keyword]",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "filter[status]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "user|favorites",
                        "required": false
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "-id, -created_at",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "posts.store",
                "operationId": "4a657e51a0066c64781851f2fd2651e2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content",
                                    "media_ids"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string"
                                    },
                                    "media_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "posts.show",
                "operationId": "b0f9d92879218f68a376b6edc6467c67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "user|favorites",
                        "required": false
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "-id, -created_at",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "posts.update",
                "operationId": "929614537d76a081a92170a41ed3bdb3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content",
                                    "media_ids"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string"
                                    },
                                    "media_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "posts.destroy",
                "operationId": "fad1726bd24cd459fc1525b67985df39",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Bearer {token}"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}